requirements.txt
bleak==0.22.2
pycryptodome==3.20.0

